Advertisement Prediction
In this project we study the various modes of advertisement which is used for a particular product and then acknowledge the change in the sales of the product. In this i have used Linear Regression to compute the results from the dataset and compared the actual and predicted values obtained by running the model.